Preparing ValuesSourceAggregatorFactory/Parser for refactoring
This change adds AbstractValuesSourceParser which will be the new class used to create ValuesSourceAggregatorFactory objects. AbstractValuesSourceParser parses all the parameters required for ValuesSource and passes to the sub-class to parse any other (implementation specific) parameters. After parsing is complete it will call createFactory on the implementing class to create the AggregatorFactory object and then set the ValuesSource specific parameters before returning it.

ValuesSourceAggregatorFactory also now has setter methods so that it can be used as the 'builder' object in the future.